requirements.txt
fastapi
uvicorn
google-cloud-speech
google-cloud-texttospeech
openai
langdetect
pytest
